+++ modulo-motores/modulo-controlador-motores/lib/motor_driver.cpp
@@ -1,5 +1,5 @@
 #include <Arduino.h>
-#include "motores_interface.h"
+#include "motor_driver.h"
 
 // Motor A
 #define PWM_MOTOR_A 17 // TX2
@@ -16,11 +16,12 @@
 #define PWM_CANAL_B 1  // Canal PWM para el motor B
 #define PWM_FRECUENCIA 5000  // Frecuencia PWM en Hz
 #define PWM_RESOLUCION 8  // Resolucion PWM (8 bits, valores de 0 a 255)
-#define VELOCIDAD_MAXIMA 255
-#define VELOCIDAD_GIRO 210
-#define VELOCIDAD_DETENIDO 0
 
-int motores_setup() {
+void setup(void) {}
+
+void loop(void) {}
+
+void motor_driver_setup() {
 
     // Configuración motor A
     pinMode(INPUT_1_MOTOR_A, OUTPUT);
@@ -39,36 +40,8 @@ int motores_setup() {
     // Configuración canales PWM
     ledcSetup(PWM_CANAL_A, PWM_FRECUENCIA, PWM_RESOLUCION);
     ledcSetup(PWM_CANAL_B, PWM_FRECUENCIA, PWM_RESOLUCION);
-
-    return 1;
 }
 
-// Módulo de movimiento
-void robot_avanzar() {
-
-    motorA_acelerar(VELOCIDAD_MAXIMA);
-    motorB_acelerar(VELOCIDAD_MAXIMA);
-}
-
-void robot_detenerse() {
-
-    motorA_acelerar(VELOCIDAD_DETENIDO);
-    motorB_acelerar(VELOCIDAD_DETENIDO);
-}
-
-void robot_girar_izquierda() {
-
-    motorA_acelerar(VELOCIDAD_DETENIDO);
-    motorB_acelerar(VELOCIDAD_GIRO);
-}
-
-void robot_girar_derecha() {
-
-    motorA_acelerar(VELOCIDAD_GIRO);
-    motorB_acelerar(VELOCIDAD_DETENIDO);
-}
-
-// Módulo de motores
 void motorA_acelerar(int intensidad) {
 
     digitalWrite(INPUT_1_MOTOR_A, HIGH);
