+++ modulo-motores/modulo-controlador-motores/lib/move_controller.cpp
@@ -1,0 +1,39 @@
+#include <Arduino.h>
+#include "move_controller.h"
+#include "motor_driver.h"
+
+// Velocidades para el movimiento de los motores
+#define VELOCIDAD_MAXIMA 255
+#define VELOCIDAD_GIRO 210
+#define VELOCIDAD_DETENIDO 0
+
+void setup(void) {
+
+    motor_driver_setup();
+}
+
+void loop(void) {}
+
+void robot_avanzar() {
+
+    motorA_acelerar(VELOCIDAD_MAXIMA);
+    motorB_acelerar(VELOCIDAD_MAXIMA);
+}
+
+void robot_detenerse() {
+
+    motorA_acelerar(VELOCIDAD_DETENIDO);
+    motorB_acelerar(VELOCIDAD_DETENIDO);
+}
+
+void robot_girar_izquierda() {
+
+    motorA_acelerar(VELOCIDAD_DETENIDO);
+    motorB_acelerar(VELOCIDAD_GIRO);
+}
+
+void robot_girar_derecha() {
+
+    motorA_acelerar(VELOCIDAD_GIRO);
+    motorB_acelerar(VELOCIDAD_DETENIDO);
+}
