+++ modulo-motores/modulo-controlador-motores/lib/move_controller.h
@@ -1,0 +1,10 @@
+#ifndef MOVE_CONTROLLER_H
+#define MOVE_CONTROLLER_H
+
+// Movimientos del robot
+void robot_avanzar();
+void robot_detenerse();
+void robot_girar_izquierda();
+void robot_girar_derecha();
+
+#endif
